framework-comparisons/03-rag-implementation/google-adk/requirements.txt
google-generativeai==0.8.3
python-dotenv==1.2.1
langchain-openai==1.1.7
langchain-community==0.4.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
